Add files via upload
Theater Application. The application prompts the user for the name of 
a text file that contains the reserved seats in a theater. It reads the text file and 
computes the number of reserved seats and the revenue. It also presents a 
theater seating chart with all the reserved seats marked with x.  
Each line in the text file contains the row and the column of a reserved seat. Each 
reservation costs $20 and the theater size is 9 rows X 10 columns. The program 
output should be formatted as shown in the Sample Run.